requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
black==25.1.0
click==8.2.1
Django==5.2.4
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
propcache==0.3.2
PyJWT==2.9.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.1
yarl==1.20.1
